Trying to figure out everything I need to code
I want to make a "Whichever Enigma variant you wanna play with" python "calculator" that is easy to work with. Ideal input is run the thing, it pops up asking for the daily settings, or if you want to reuse settings from the previous session (Don't know if these should save at end of message, or reset to beginning of day settings) 

Then you configure the machine, which rotors are used, and in which order, how many, which ones have advancing turned on, and at which rate - Ideally I want there to be historically accurate templates - and I think I want those to be csv files, mostly because I'm familiar with them.

Then it should ask for plug board (Steckerbrett/steckerboard) settings. 

Then you should be off to the races.

Type in your message and it should export as enciphered text broken into five letter all caps blocks printed to console - or returned to another function/process/variable.

The plan to handle the encoding is to take the user input, and I haven't even thought about the algorithmic complexity of this yet.

Okay - I have to take each character (take string, drop punctuation, allcaps, 5 block spacing) and apply the modulo 26 addition that each rotor of steckerboard entry does to it. it might be easier to think of the letters as numbers, and the values of the rotors as numerical increments rather than alphabetic substitutions. 

Dangit. it took such a long time to get that formatted that nicely.

Okay, wait a second - I COULD use the previous letter entered 

user_input = ABE LINCOLN
A maps to a variable lets go with 
user_input [1]
since it IS an A we have to look at what is in the A wiring position on the input side of the first Rotor, which we determined at startup. because the A wiring position is absolute - but the position of the rotor is relative, and I've decided to store the rotor as a list - it'd be best to shift the index of A's wiring - that's going to be the tricky bit, because each of the rotors shifts the value of our variable (Wheel 1 for instance is in a position where A is in position A, so index is not shifted for this wheel, so A is encoded on wheel IC as D, so now we exit at index 3 - to the next wheel.

So now I need to track that, index 0 to index 3, and A to D.

I can't really use a hash table because if I change both the key and the value everything becomes meaningless because they have no relative position. So I've got to generate a second array to track the position of our "Electrical signal" to it's new index - only now this wheel has been rotated so we now need to add (Modulo 26) 13 to any index coming in here because wheel number 2 was advanced 13 places, (And in 632 letters will advance 1 more, maybe 2 more depending on initial setup) so the function that handles that has to take in a lot of variables from the initial setup. to handle this properly.

I am wondering if it is possible to handle seemingly discrete sections of the encryption discretely - but still do not know enough about how the plug board is wired to the rest of the machine to know if I can separate those two processes totally. 

I really feel like I should be using a dict/hash table for something but I just don't think I know the right way to use them yet. On the up side I think the ways in which I am trying to use lists once the program is running should be decent.

I don't want to make all the lists in the above code - I think what I need to do is put those into flow control statements in a function so that the specific rotor you want goes into the slot you want, in the position you want. If II can figure out how to decrement or increment the index of the list and make it wrap without it taking stupid amounts of memory (I mean, the most rotors is going to be four, and the lists are only 26 chars long) 

Okay, so the rotors go into x new variables "rotor_a" "rotor_b" "rotor_c" "rotor_d" etc. if more are enabled later...

so, I take my A turned D, and I apply it to the second wheel, this time because it's at index 16 effectively it enters "p" Well, Lucky us - it comes out D again because we have wheel IIC in position 2. so, user_input[1] stays D

BUT instead of being shot to index 3, we're now at index 16 - that one was easy. it usually won't be that easy, but it is our new absolute letter index, plus our rotation index modifier. 

now the next wheel should be a reflector. 

and it should be reflector a, so lets say it was advanced 4 positions, so we'll have to enter it at index 20. that puts us at S which spits us out at T, that's funny. and returns us to index 19 on the encrypted side of wheel iic

So now user_input[1] = T and tracking_position[1] = 19

so  now we reverse the course through the wheels offset apparently by one slot because of the reflector. and we report a letter at the end of the path for user_input[1]

then some of the letters are swapped with the plug board.

Okay, that can be coded discretely. 

Okay, this is going to take a little while longer to plan, but I think I have a good skeleton here. and I think I'm going to be able to read it and figure out whether I was being an idiot or not. 

I'll try to keep this updated as I go, and try to make it more into real documentation, and less like a live internal dialogue.